Refactor av api-handlere
